requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.1
pillow==10.4.0
proglog==0.1.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
tqdm==4.66.5
urllib3==2.2.3
